refactor data ms corresponding to sql script

+++ src/main/java/priam/data/priamdataservice/entities/Processing.java
@@ -22,7 +22,7 @@ public class Processing {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long processingId;
+    private int processingId;
 
     private String processingName;
